+++ CHANGELOG.md
@@ -11,6 +11,7 @@ colorscad:
 ### Fixed
 colorscad: improved error handling
 - Input file does not exist
+- The wrong 'sort' command is used
 
 ## [0.3.1] - 2021-07-03
 ### Fixed